Rework stack traces.
Refactor setup and calling code, separate into separate source file.
mtev_main will now setup sefl-diagnosis for foreground operated processes.
Add SIGIOT handling if it is present.
