+++ src/Makefile.in
@@ -110,6 +110,7 @@ MAPPEDHEADERS=utils/mtev_atomic.h utils/mtev_b32.h utils/mtev_b64.h \
    utils/mtev_watchdog.h utils/mtev_uuid_parse.h                    \
    utils/mtev_perftimer.h utils/mtev_zipkin.h                       \
    utils/mtev_hyperloglog.h json-lib/mtev_arraylist.h               \
+   utils/mtev_stacktrace.h                                          \
    json-lib/mtev_bits.h json-lib/mtev_debug.h                       \
    json-lib/mtev_json_object.h json-lib/mtev_json_tokener.h         \
    json-lib/mtev_json_util.h json-lib/mtev_json.h                   \
@@ -144,7 +145,9 @@ MTEV_UTILS_OBJS=utils/mtev_b32.hlo utils/mtev_b64.hlo                 \
   utils/mtev_skiplist.hlo utils/mtev_sort.hlo utils/mtev_str.lo       \
   utils/mtev_watchdog.lo utils/mtev_zipkin.lo utils/mtev_memory.lo    \
   utils/mtev_cht.hlo utils/mtev_uuid_parse.hlo                        \
-  utils/mtev_perftimer.lo utils/mtev_hyperloglog.hlo $(ATOMIC_OBJS)
+  utils/mtev_perftimer.lo utils/mtev_hyperloglog.hlo                  \
+  utils/mtev_stacktrace.lo
+  $(ATOMIC_OBJS)
 
 LIBMTEV_OBJS=mtev_main.lo mtev_listener.lo mtev_cluster.lo            \
    mtev_console.lo mtev_console_state.lo mtev_console_telnet.lo       \
